requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
openai==1.12.0
requests==2.31.0
pyjwt==2.8.0
python-dotenv==1.0.0
email-validator==2.1.0
werkzeug==2.3.7
azure-cosmos==4.5.1
azure-search-documents==11.4.0